requirements.txt
pandas
tensorflow
kafka-python
pyspark
prometheus-client